requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
distro==1.9.0
Django==4.2.7
exceptiongroup==1.3.0
httpx==0.24.1
idna==3.10
jiter==0.10.0
numpy==2.0.2
openai==1.3.5
opencv-python==4.12.0.88
Pillow==10.1.0
pydantic==2.11.9
pydantic_core==2.33.2
python-decouple==3.8
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0

# Voice Recognition
pvporcupine>=3.0.0
PyAudio>=0.2.14
python-dotenv>=1.0.0

# Django REST Framework
djangorestframework==3.14.0

# GPT-SoVITS TTS (Mock - 실제 사용 시 GPT-SoVITS 설치 필요)
# torch==2.1.2
# torchaudio==2.1.2
# librosa==0.10.1
# soundfile==0.12.1
# scipy==1.11.4
